feat: Add drink creation and display

+++ app/src/main/java/com/example/poputka/common/global_state/local_settings_state/SettingsState.kt
@@ -3,7 +3,7 @@ package com.example.poputka.common.global_state.local_settings_state
 import com.example.poputka.common.domain.model.VolumeUnit
 
 data class AppPreferencesState(
-    val goalSetting: Double = 1500.0,
+    val goalSetting: Int = 1500,
     val volumeUnitSetting: VolumeUnit = VolumeUnit.Milliliters,
     val autoCalculation: Boolean = true
 )
